+++ frontend/src/views/AllPeriodInfo.vue
@@ -40,7 +40,7 @@ setup() {
     try{
         const url = process.env.VUE_APP_BACKEND_URL + 'activities/total';
         const response = await axios.get(url,
-                                        {headers: {Authorization: `${store.state.tokenType} ${store.state.accessToken}`}})
+                                        {headers: {Authorization: `${store.state.authenticateModule["tokenType"]} ${store.state.authenticateModule["accessToken"]}`}})
         if (response.status===200){
             message.value = [`合計:${response.data.total_income}万円\n`,
                             `内訳\n`,
